# docs/swagger.yaml
basePath: /
definitions:
...
  handler.ApproveRejectPerizinanInput:
    properties:
      keterangan_guru:
        example: Izin disetujui, semoga lekas sembuh.
        type: string
      status:
        example: Disetujui
        type: string
    type: object
...
  helpers.SwaggerError400Response:
    properties:
      message:
        example: Input tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError401Response:
    properties:
      message:
        example: Token tidak ditemukan atau tidak valid
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError403Response:
    properties:
      code:
        example: ERR_PERMISSION_DENIED
        type: string
      message:
        example: 'Akses ditolak: Anda tidak memiliki akses untuk fitur ini'
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError404Response:
    properties:
      message:
        example: Data tidak ditemukan
        type: string
      success:
        example: false
        type: boolean
    type: object
  helpers.SwaggerError500Response:
    properties:
      message:
        example: Terjadi kesalahan internal pada server
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  model.Guru:
    properties:
      alamat:
        description: Deprecated Database Columns (Retained for Non-Destructive Migrations)
        example: Jl. Diponegoro No. 25
        type: string
      alamat_detail:
        example: Jl. Diponegoro No. 25
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      desa:
        example: Sidorejo Lor
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      jenis_kelamin:
        example: Laki-laki
        type: string
      kabupaten:
        example: Salatiga
        type: string
      kecamatan:
        example: Sidorejo
        type: string
      mapel_ids:
        description: Virtual field for frontend multiselect checkboxes
        example:
        - 1
        - 2
        items:
          type: integer
        type: array
      nama:
        example: Ahmad Guru
        type: string
      nip:
        example: GR001
        type: string
      no_hp:
        example: "081234567890"
        type: string
      photo_url:
        example: /uploads/profile/guru_1.jpg
        type: string
      provinsi:
        description: Address Normalization Fields
        example: Jawa Tengah
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        example: 3
        type: integer
    type: object
...
  model.Kelas:
    properties:
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      jurusan:
        example: Rekayasa Perangkat Lunak
        type: string
      kapasitas:
        example: 36
        type: integer
      nama_kelas:
        example: X RPL 1
        type: string
      tingkat:
        example: X
        type: string
      total_siswa:
        example: 32
        type: integer
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      wali_kelas:
        $ref: '#/definitions/model.Guru'
      wali_kelas_id:
        example: 1
        type: integer
    type: object
...
  model.Perizinan:
    properties:
      alasan:
        example: Ada keperluan keluarga
        type: string
      attachment_url:
        example: /uploads/perizinan/sample.pdf
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      disetujui_oleh:
        example: 1
        type: integer
      guru:
        $ref: '#/definitions/model.Guru'
      id:
        example: 1
        type: integer
      keterangan_guru:
        example: Semoga lekas sembuh/urusan lancar
        type: string
      siswa:
        $ref: '#/definitions/model.Siswa'
      siswa_id:
        example: 1
        type: integer
      status:
        description: '"Pending", "Disetujui", "Ditolak"'
        example: Pending
        type: string
      tanggal_mulai:
        example: "2026-07-15T00:00:00Z"
        type: string
      tanggal_selesai:
        example: "2026-07-16T00:00:00Z"
        type: string
      tipe:
        description: '"Izin" atau "Sakit"'
        example: Izin
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
  model.PerizinanResponseList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Perizinan'
        type: array
      message:
        example: Berhasil mengambil data perizinan
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.PerizinanResponseSingle:
    properties:
      data:
        $ref: '#/definitions/model.Perizinan'
      message:
        example: Berhasil mengambil detail perizinan
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  model.SchoolProfile:
    properties:
      academic_status:
        example: "\U0001F7E2 Aktif"
        type: string
      academic_year:
        example: 2026/2027
        type: string
      accreditation:
        example: A (Sangat Baik)
        type: string
      address:
        example: Jl. Diponegoro No. 25, Salatiga
        type: string
      appointment_period:
        example: 2021 - 2027
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      current_semester:
        example: Ganjil
        type: string
      curriculum:
        example: Kurikulum Merdeka
        type: string
      email:
        example: info@smkn1salatiga.sch.id
        type: string
      established_year:
        example: "1965"
        type: string
      facilities:
        example: Perpustakaan, Lab Komputer, Bengkel Kerja, Lapangan Olahraga
        type: string
      id:
        example: 1
        type: integer
      level:
        example: SMK / Sekolah Menengah Kejuruan
        type: string
      logo:
        example: /uploads/profile/logo.png
        type: string
      mission:
        example: |-
          1. Menyelenggarakan pembelajaran berkualitas...
          2. Membina budi pekerti luhur...
        type: string
      name:
        example: SMK Negeri 1 Salatiga
        type: string
      npsn:
        example: "20312345"
        type: string
      operational_status:
        example: Aktif
        type: string
      phone:
        example: (0298) 123456
        type: string
      postal_code:
        example: "50711"
        type: string
      principal_name:
        example: Drs. Hadi Santoso
        type: string
      principal_nip:
        example: "197205121998031002"
        type: string
      principal_position:
        example: Kepala Sekolah
        type: string
      school_type:
        example: Sekolah Menengah Kejuruan (SMK)
        type: string
      shift:
        example: Pagi
        type: string
      status:
        example: Negeri
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      vision:
        example: Menjadi lembaga pendidikan kejuruan yang unggul, berkarakter, dan
          berdaya saing global.
        type: string
      website:
        example: https://smkn1salatiga.sch.id
        type: string
    type: object
  model.SchoolResponseSingle:
    properties:
      data:
        $ref: '#/definitions/model.SchoolProfile'
      message:
        example: Berhasil mengambil profil sekolah
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.Siswa:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25
        type: string
      alamat_detail:
        example: Jl. Diponegoro No. 25
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      desa:
        example: Sidorejo Lor
        type: string
      email:
        example: budi.santoso@sekolah.com
        type: string
      id:
        example: 1
        type: integer
      jenis_kelamin:
        example: Laki-laki
        type: string
      kabupaten:
        example: Salatiga
        type: string
      kecamatan:
        example: Sidorejo
        type: string
      kelas:
        $ref: '#/definitions/model.Kelas'
      kelas_id:
        example: 1
        type: integer
      nama:
        example: Budi Santoso
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      photo_url:
        example: /uploads/profile/siswa_1.jpg
        type: string
      provinsi:
        description: Address Normalization Fields
        example: Jawa Tengah
        type: string
      tanggal_lahir:
        example: "2010-05-15"
        type: string
      tempat_lahir:
        description: Deprecated Database Columns (Retained for Non-Destructive Migrations)
        example: Salatiga
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      user_id:
        example: 2
        type: integer
    type: object
...
  model.User:
    properties:
      alamat:
        example: Jl. Diponegoro No. 25, Salatiga
        type: string
      created_at:
        example: "2026-07-15T19:00:00Z"
        type: string
      email:
        example: ahmad.guru@sekolah.com
        type: string
      gelar:
        example: S.Pd.
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_first_login:
        example: true
        type: boolean
      jenis_kelamin:
        example: Laki-laki
        type: string
      last_login_at:
        type: string
      nama:
        example: Ahmad Guru
        type: string
      nip:
        description: Ignored by GORM, populated for unified API login response
        example: GR001
        type: string
      nis:
        example: "12345678"
        type: string
      no_hp:
        example: "081234567890"
        type: string
      role:
        example: guru
        type: string
      updated_at:
        example: "2026-07-15T19:00:00Z"
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: praktikum@example.com
    name: Praktikum Pemrograman III
  description: Dokumentasi API backend Sistem Informasi Sekolah menggunakan Golang
    Fiber, GORM, PostgreSQL, dan JWT.
  title: API Sistem Informasi Sekolah
  version: "1.0"
paths:
...
  /api/perizinan/:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh data perizinan yang relevan bagi user yang sedang
        login (siswa hanya melihat miliknya, guru melihat data kelasnya/seluruhnya).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PerizinanResponseList'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil data perizinan
      tags:
      - Perizinan
    post:
      consumes:
      - application/json
      description: Mengajukan perizinan baru (misal izin sakit/keperluan keluarga)
        untuk siswa.
      parameters:
      - description: Payload perizinan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Perizinan'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PerizinanResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ajukan permohonan perizinan
      tags:
      - Perizinan
  /api/perizinan/{id}/approve:
    put:
      consumes:
      - application/json
      description: Mengubah status persetujuan perizinan siswa. Hanya dapat dilakukan
        oleh Guru.
      parameters:
      - description: ID Perizinan
        in: path
        name: id
        required: true
        type: integer
      - description: Payload persetujuan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ApproveRejectPerizinanInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PerizinanResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.SwaggerError404Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Setujui atau tolak permohonan perizinan
      tags:
      - Perizinan
...
  /api/school-profile/:
    get:
      consumes:
      - application/json
      description: Mengambil informasi profil sekolah (ID=1). Jika belum ada di database,
        akan mengembalikan profil default.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SchoolResponseSingle'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Ambil profil sekolah
      tags:
      - School Profile
    put:
      consumes:
      - application/json
      description: Memperbarui data profil sekolah (ID=1). Hanya dapat diakses oleh
        Admin.
      parameters:
      - description: Payload profil sekolah
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SchoolProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SchoolResponseSingle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.SwaggerError400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.SwaggerError401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.SwaggerError403Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.SwaggerError500Response'
      security:
      - BearerAuth: []
      summary: Perbarui profil sekolah
      tags:
      - School Profile
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"